refactor: extract ToolListManager from McpProxy
Extract tool caching and list operations into ToolListManager:
- GetCachedTools(), PersistToolCacheIfNeeded(), RefreshCachedToolsFromUpstreamAsync()
- ListToolsWithTimeoutAsync(), FetchToolsFromUpstreamAsync(), AppendBuiltInTools()
- InitializeToolCachePath(), MarkShouldRefresh()

HealthService now uses ToolListManager.CachedToolCount directly.
